# packages/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "DevelopmentOperatorLoginDto": {
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 320,
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 12,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "GithubLoginStartDto": {
        "properties": {
          "authorizationUrl": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "authorizationUrl"
        ],
        "type": "object"
      },
...
      "OperatorSessionDto": {
        "properties": {
          "csrfToken": {
            "type": "string"
          },
          "operator": {
            "additionalProperties": false,
            "properties": {
              "authenticationMethod": {
                "enum": [
                  "DEVELOPMENT_PASSWORD",
                  "GITHUB"
                ],
                "type": "string"
              },
              "nodeIds": {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              "operatorId": {
                "format": "uuid",
                "type": "string"
              },
              "role": {
                "enum": [
                  "ADMIN",
                  "OPERATOR",
                  "VIEWER"
                ],
                "type": "string"
              }
            },
            "required": [
              "operatorId",
              "role",
              "nodeIds",
              "authenticationMethod"
            ],
            "type": "object"
          }
        },
        "required": [
          "csrfToken"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      },
      "operatorSession": {
        "description": "Environment-specific HttpOnly operator session cookie; development uses a local cookie name.",
        "in": "cookie",
        "name": "__Host-taven_admin",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Canonical HTTP contract for Taven clients",
    "title": "Taven API",
    "version": "0.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/admin/auth/github/callback": {
      "get": {
        "operationId": "OperatorAuthController_completeGithub",
        "parameters": [
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {}
          },
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "maxLength": 2048,
              "minLength": 20
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "maxLength": 2048,
              "minLength": 20
            }
          }
        ],
        "responses": {
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        },
        "summary": "Complete the browser-bound GitHub operator authorization callback",
        "tags": [
          "operator authentication"
        ]
      }
    },
    "/admin/auth/github/start": {
      "post": {
        "operationId": "OperatorAuthController_startGithub",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubLoginStartDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "description": ""
          },
          "429": {
            "description": ""
          }
        },
        "summary": "Start the production or staging GitHub operator authorization flow",
        "tags": [
          "operator authentication"
        ]
      }
    },
    "/admin/auth/login": {
      "post": {
        "operationId": "OperatorAuthController_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevelopmentOperatorLoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorSessionDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          },
          "429": {
            "description": ""
          }
        },
        "summary": "Create a development-only operator session from a provisioned password",
        "tags": [
          "operator authentication"
        ]
      }
    },
...
    "/admin/auth/session": {
      "delete": {
        "operationId": "OperatorAuthController_logout",
        "parameters": [
          {
            "description": "Required for unsafe operator requests. Obtain the session-bound value from GET /admin/auth/session.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Revoke the current operator session",
        "tags": [
          "operator authentication"
        ]
      },
      "get": {
        "operationId": "OperatorAuthController_session",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorSessionDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "operatorSession": []
          }
        ],
        "summary": "Read the current authenticated operator session and CSRF token",
        "tags": [
          "operator authentication"
        ]
      }
    },
...
  },
  "servers": [],
...
}